Rename consensus directory to raftex
