+++ CMakeLists.txt
@@ -105,10 +105,10 @@ add_subdirectory(interface)
 add_subdirectory(common)
 add_subdirectory(dataman)
 add_subdirectory(client)
-#add_subdirectory(storage)
 add_subdirectory(server)
 add_subdirectory(console)
-add_subdirectory(consensus)
+add_subdirectory(raftex)
+#add_subdirectory(storage)
 
 add_dependencies(common third-party)
 #add_dependencies(storage_engines common)
