--- consensus/Replica.cpp
@@ -1,17 +1,0 @@
-/* Copyright (c) 2018 - present, VE Software Inc. All rights reserved
- *
- * This source code is licensed under Apache 2.0 License
- *  (found in the LICENSE.Apache file in the root directory)
- */
-
-#include "consensus/Replica.h"
-
-namespace vesoft {
-namespace vgraph {
-namespace consensus {
-
-
-}  // namespace consensus
-}  // namespace vgraph
-}  // namespace vesoft
-
