+++ raftex/CMakeLists.txt
@@ -1,0 +1,8 @@
+add_library(
+    raftex_obj OBJECT
+    FileBasedWal.cpp
+)
+add_dependencies(raftex_obj common)
+
+add_subdirectory(test)
+
